git-svn-id: http://svn.code.sf.net/p/itsucks/code/trunk@432 540001c9-3a26-0410-bb9e-87544be22eba

+++ itsucks-core/src/main/java/de/phleisch/app/itsucks/processing/download/impl/PersistenceProcessor.java
@@ -205,7 +205,7 @@ public class PersistenceProcessor extends AbstractDataProcessor implements DataP
 		}
 		mBufferedOut = null;
 		
-		if(mFile != null && pDeleteFile) {
+		if(mFile != null && mFile.exists() && pDeleteFile) {
 			boolean deleteSucessful = mFile.delete();
 			if(!deleteSucessful) {
 				mLog.error("Could not delete partial file: " + mFile.getAbsolutePath());
